refactor: removed `command` option BREAKING CHANGE: Removed `command` option in favour of plugin enforcement
To reduce any future complexity, you can define when the ES import rewriter will be executed. This can be done as shown in the below example by setting the `apply` option to either `build` or `serve`.

Ensure that `enforce` is always set to `pre` as it will prevent Vite from transforming things first.

```js
plugins: [
  {
    ...aemViteImportRewriter({ /* ... */ }),

    apply: 'build',
    enforce: 'pre',
  },
]
```